Move users/api contents a level down
